+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [8.10.3]
+
 ## [8.10.2]
 
 ### Fixed
@@ -13736,7 +13738,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#957](https://github.com/MetaMask/metamask-mobile/pull/957): fix timeouts (#957)
 - [#954](https://github.com/MetaMask/metamask-mobile/pull/954): Bugfix: onboarding navigation (#954)
 
-[Unreleased]: https://github.com/MetaMask/metamask-mobile/compare/v8.10.2...HEAD
+[Unreleased]: https://github.com/MetaMask/metamask-mobile/compare/v8.10.3...HEAD
+[8.10.3]: https://github.com/MetaMask/metamask-mobile/compare/v8.10.2...v8.10.3
 [8.10.2]: https://github.com/MetaMask/metamask-mobile/compare/v8.10.1...v8.10.2
 [8.10.1]: https://github.com/MetaMask/metamask-mobile/compare/v8.10.0...v8.10.1
 [8.10.0]: https://github.com/MetaMask/metamask-mobile/compare/v8.9.1...v8.10.0